requirements.txt
anyio==3.7.0
appnope==0.1.3
arrow==1.2.3
asttokens==2.2.1
backcall==0.2.0
bard==0.1
bcrypt==4.0.1
binaryornot==0.4.4
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
cookiecutter==2.1.1
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
dnspython==2.3.0
executing==1.2.0
fastapi==0.99.0
future==0.18.3
google-api-core==2.11.1
google-api-python-client==2.91.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
GoogleBard==1.4.0
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httpx==0.24.1
idna==3.4
ipykernel==6.22.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jinja2-time==0.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
netmiko==4.2.0
ntc-templates==3.5.0
ordered-set==4.1.0
packaging==23.1
paramiko==3.2.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
pymongo==4.4.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyserial==3.5
python-dateutil==2.8.2
python-google-calendar-api==0.5
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.31.0
rich==13.4.2
robotframework==6.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scp==0.14.5
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
stack-data==0.6.2
starlette==0.27.0
text-unidecode==1.3
textfsm==1.1.3
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.7.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
